Error compilación y una excepción no controlada
El error en la compilación por los paquetes fue arreglado y se detecto y reparó una excepción no controlada.

+++ Proyecto1ED1/Models/Model/PatientInfo.cs
@@ -9,7 +9,7 @@ namespace Proyecto1ED1.Models
     {
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int DPI_Partida { get; set; }
+        public long DPI_Partida { get; set; }
         public string Departamento { get; set; }
         public string Municipio { get; set; }
         public string Sintomas { get; set; }
